Deployments/Amazon Seller-Order Status Prediction/requirements.txt
certifi==2021.10.8
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
itsdangerous==2.0.1
gunicorn==20.1.0
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
numpy1.21.5
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
threadpoolctl==3.1.0
Werkzeug==2.0.3
wincertstore==0.2
